Modified DatabaseSupport for list of current users.
Had to add a new column (currentlist) to Channel table of database. This
meant I had to go in and update putChannel and getChannel methods of
database support.

I also got /listChannelUsers working, but I had to debug a TON. Right
now there are an immense amount of print statements being called as I
was trying to track down my issues. So i'll have to remove those before
I merge.